LangGraph/requirements_langgraph.txt
# ─────────────────────────────────────────────
# Agentic4API — LangGraph Agent Discovery
# Version compatible Pydantic 2.10+
# ─────────────────────────────────────────────

# Orchestration agent
langgraph>=0.2.60,<0.3.0
langchain>=0.3.14,<0.4.0
langchain-core>=0.3.29,<0.4.0
langchain-mistralai>=0.2.4,<0.3.0

# Interface chat — Chainlit 2.x corrige l'incompatibilité Pydantic
# (bug "CodeSettings is not fully defined" présent en 1.3.2)
chainlit>=2.0.0,<3.0.0

# Mémoire persistante (remplace Memory Buffer N8N)
langgraph-checkpoint-sqlite>=2.0.6

# HTTP & utilitaires
requests>=2.32.3
python-dotenv>=1.0.1